requirements.txt
ase~=3.22
oteapi-core~=0.1
